Exploratory Data Analysis
In this module, begin working with the cleaned dataset from the previous module. You will now begin to analyze the dataset to find the distribution of data, presence of outliers and the correlation between different columns.

Learning Objectives:

Examine the distribution of data in the dataset.
Identify the outliers in the dataset.
Remove outliers from the dataset.
Recognize the correlation between features in the dataset.